+++ Cargo.toml
@@ -28,6 +28,7 @@ gltf = "0.11"
 
 [dependencies.rendy]
 optional = true
-# path = "../rendy/rendy"
-git = "https://github.com/omni-viral/rendy"
-rev = "04cf6a6e58c57b3a9b74cf6eeaf522f5c40194a8"
+path = "../rendy/rendy"
+features = ["texture-image"]
+# git = "https://github.com/termhn/rendy"
+# rev = "595d748c934f24be3ccbcd49343a8ea2d19257a9"
