+++ src/node/env_preprocess/mod.rs
@@ -1,0 +1,33 @@
+use rendy::{
+    command::{QueueId, FamilyId},
+    factory::ImageState,
+    texture::Texture,
+};
+
+use gfx_hal as hal;
+
+pub mod equirectangular_to_cube_faces;
+pub mod faces_to_cubemap;
+
+pub struct Aux<B: hal::Backend> {
+    align: u64,
+    equirectangular_texture: Texture<B>,
+    environment_cubemap: Texture<B>,
+}
+
+impl<B> faces_to_cubemap::FacesToCubemapResource<B> for Aux<B>
+    where B: hal::Backend
+{
+    fn get_cubemap(&self) -> &Texture<B> {
+        &self.environment_cubemap
+    }
+
+    fn cubemap_end_state(&self) -> ImageState {
+        ImageState {
+            queue: QueueId(hal::queue::QueueFamilyId(0), 0), // TODO
+            stage: hal::pso::PipelineStage::FRAGMENT_SHADER,
+            access: hal::image::Access::SHADER_READ,
+            layout: hal::image::Layout::ShaderReadOnlyOptimal,
+        }
+    }
+}
